+++ resources/views/admin/partners/index.blade.php
@@ -68,6 +68,9 @@
                         </span>
                     </a>
 
+                    @if ($partner->isNotValidated())
+                        <span style="color: #ce7a10;"> (non-validé)</span>
+                    @endif
 @php
     $currencyExchanges = [];
     foreach ($partner->locations as $location) {
@@ -118,6 +121,26 @@
                                 {{ $partner->name_sort }}
                             </span>
                         </a>
+
+                        @if ($partner->isNotValidated())
+                            <span style="color: #ce7a10;"> (non-validé)</span>
+                        @endif
+@php
+    $currencyExchanges = [];
+    foreach ($partner->locations as $location) {
+        if ($location->currencyExchange) {
+            $currencyExchanges[] =
+                'Comptoir de change';
+        }
+    }
+@endphp
+                    @if (count($currencyExchanges) === 1)
+                        <br>
+                        A un comptoir de change
+                    @elseif (count($currencyExchanges) > 1)
+                        <br>
+                        A {{ count($currencyExchanges) }} comptoirs de change
+                    @endif
                     </li>
                     @endif
                 @endforeach